requirements.txt
fastapi
uvicorn
langchain_groq
PyGithub
gitpython
python-dotenv
celery
redis
requests
